requirements.txt
alembic==1.1.0
ecdsa==0.13.2
falcon==2.0.0
falcon-auth==1.1.0
gunicorn==19.9.0
Mako==1.1.0
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.3
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
uuid==1.30
voluptuous==0.11.7
